Create tone generator
Use hard-coded frequency of 440 Hz, which corresponds to an A.
It's a reference frequency commonly used for calibration purposes.

Embracing static-site-generator-webpack-plugin's default behaviour
which results in paths such as http://jelefra.com/tone-generator/
(includes a trailing slash). This path structure is similar to the one
used by Netlify and seems common among static sites.

A workaround to enable http://jelefra.com/tone-generator (without the
trailing slash) would require to specify the path as
'/tone-generator.html' in the webpack config and to rely on Netlify to
strip out the .html. An extra step would be needed to fix routing in
the dev environment.

See https://developer.mozilla.org/en-US/docs/Web/API/Web_Audio_API/Simple_synth