Add React Helmet
Ensure the <head> (especially the page title) updates on the client
by adding React Helmet.

As Helmet.renderStatic() needs to be called after
ReactDOMServer.renderToString, this change requires the structure of
the app to be modified: the Root component is replaced with the html
template and the custom locals are no longer used by the static plugin.

See https://github.com/nfl/react-helmet