Create AudioContext after a user interaction
Even though the oscillator is started after the user clicks the "Play"
button, Chrome warns about the annoyance of automatically playing
audio.

Chrome's behaviour is inconsistent and the warning message is
not always logged.

Chrome error: "The AudioContext was not allowed to start. It must be
resumed (or created) after a user gesture on the page.
https://goo.gl/7K7WLu"