Add favicon
Get rid of the console error "GET https://jelefra.com/favicon.ico 404".

Other images considered:
  puzzle piece
  olive branch
  fire
  compass
  frog
  hot pepper

Chose the smiling blob as it's simple, works in light and dark mode,
and fits  a personal website.

Happens to be the main example on https://en.wikipedia.org/wiki/Emoji

See https://evilmartians.com/chronicles/how-to-favicon-in-2021-six-files-that-fit-most-needs