Set frequency to a given note
With A4 = 440 Hz.

Tempted to replace:
  const octaves = [0, 1, 2, 3, 4, 5, 6, 7, 8];
With:
  const octaves = Array.from(Array(9).keys());

But simplicity and readability trump cleverness.

Wanted to use grid-column-gap but since grid-row-gap can't be used (as
each grid is one row only), spacing is managed via a margin set on the
buttons instead, for consistency.

Unsure whether it is better to use 'display: none' or to remove the
modal from the DOM when the modal is closed. Went for the former.

See:
  https://pages.mtu.edu/~suits/notefreqs.html
  https://www.digitalocean.com/community/tutorials/react-modal-component
  https://css-tricks.com/considerations-styling-modal/