Automate image manipulation
Update the size of mobile images from 750*469px to 768*480px to
maintain the exact aspect ratio of 16:10.

End up using the same image size (768*480px) for mobile and standard
desktop displays given mobile devices have a pixel density of
at least 2 and a width around 375px (375 * 2 ~= 768).

It is slightly beneficial to zoom into the image on mobile to highlight
the subject. Using the same image as on standard desktop displays
would be acceptable and simpler.

See https://sharp.pixelplumbing.com/api-resize